requirements.txt
fastapi
uvicorn
joblib
pandas
dvc[gcs]
scikit-learn
pydantic
google-cloud-storage 
opentelemetry-exporter-gcp-trace
